Adds a way to filter CfPs
This change was mainly made to easily add more filters for CfPs. The
specific need was a filter that removes parameters from the URL of an
event to make the URI unique but still identify the event. Up to now
some event-uris contained a path and where therefore not considered the
same for the same events. That leads to accumulation of events due to
different Cookie-IDs and such…